Stop an empty report cell fataling every summed report
Utils_RecordBrowser_Reports sums each column by adding cell values onto an int
accumulator: $total[$k] += strip_tags($w). A cell is a display string, so it can
be non-numeric - most often simply empty, for a row with no value in that
column.

PHP 8's saner string-to-number conversions turned "int + non-numeric string" into
a TypeError, and '' counts as non-numeric. So a single empty cell in a summed
column now takes down the whole report with "Unsupported operand types: int +
string", where PHP 7 read it as 0 and carried on. Reported against Warehouse's
Sales Report -> value by warehouse, but nothing about it is Warehouse-specific:
any report with a row or column summary and one gap in the data hits it.

Casting to float restores exactly the old coercion (leading numeric prefix, else
0) and cannot throw. Verified across the shapes a cell can hold - '', '&nbsp;',
'abc', '12.5 PLN', '123' - and the resulting float still renders without a
trailing ".0", so no visible change to correct reports. It is also already the
idiom in this file: the chart code below does (float)strip_tags() at five places
for the same reason.

Fixed at all five summing sites, not just the two in the reported trace - the
others sit in the no-categories branch and the CSV/PDF path, and fail the same
way on a differently shaped report.

+++ modules/Utils/RecordBrowser/Reports/Reports_0.php
@@ -414,14 +414,25 @@ class Utils_RecordBrowser_Reports extends Module {
 					if ($this->row_summary!==false) {
 						foreach ($res_ref as $k=>$w) {
 							if (!isset($total[$k])) $total[$k] = 0;
-							$total[$k] += strip_tags($w);
+							// (float) is load-bearing, not decoration: a report cell is a
+							// display string, so it can be non-numeric - most commonly just
+							// empty for a row with no value in that column. PHP 8's saner
+							// string-to-number conversion made "int + non-numeric string" a
+							// TypeError ("Unsupported operand types: int + string"), and ''
+							// counts as non-numeric, so one empty cell in a summed column
+							// fatals the whole report where PHP 7 silently read it as 0.
+							// The cast restores exactly the old coercion - leading numeric
+							// prefix, else 0 - and cannot throw. Same idiom as the
+							// (float)strip_tags() calls in the chart code further down.
+							// Repeated at all five summing sites; full note here only.
+							$total[$k] += (float)strip_tags($w);
 						}
 					}
 					if ($this->col_summary!==false) {
 						if (!isset($this->cols_total[$i])) $this->cols_total[$i] = array();
 						foreach ($res_ref as $k=>$w) {
 							if (!isset($this->cols_total[$i][$k])) $this->cols_total[$i][$k] = 0;
-							$this->cols_total[$i][$k] += strip_tags($w);
+							$this->cols_total[$i][$k] += (float)strip_tags($w);
 						}
 					}
 					$res_ref = $this->format_cell($this->format, $res_ref);
@@ -458,14 +469,14 @@ class Utils_RecordBrowser_Reports extends Module {
 							foreach ($v[$c] as $k=>$w) {
 								if (!isset($total[$k])) $total[$k] = 0;
 								if($w) $empty = false;
-								$total[$k] += strip_tags($w);
+								$total[$k] += (float)strip_tags($w);
 							}
 						}
 						if ($this->col_summary!==false) {
 							if (!isset($this->cols_total[$c][$i])) $this->cols_total[$c][$i] = array();
 							foreach ($v[$c] as $k=>$w) {
 								if (!isset($this->cols_total[$c][$i][$k])) $this->cols_total[$c][$i][$k] = 0;
-								$this->cols_total[$c][$i][$k] += strip_tags($w);
+								$this->cols_total[$c][$i][$k] += (float)strip_tags($w);
 							}
 						}
 						$next = $this->format_cell($format, $v[$c]);
@@ -536,7 +547,7 @@ class Utils_RecordBrowser_Reports extends Module {
 					if ($this->row_summary!==false) {
 						foreach ($res_ref as $k=>$w) {
 							if (!isset($total[$k])) $total[$k] = 0;
-							$total[$k] += strip_tags($w);
+							$total[$k] += (float)strip_tags($w);
 						}
 					}
 					$res_ref = $this->format_cell(array($this->format,'total'), $res_ref, 'col_total');
